+++ CHANGELOG.md
@@ -19,6 +19,6 @@ All notable changes to this project will be documented in this file.
   
 ## v0.1.0-alpha
 ### What's Changed
-  - Feat/laptop desktop by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/1
+  - Feat/laptop desktop by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/2
   - fix: wrong way to write grub file by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/2
   - Initial setup script for Fedora